requirements.txt
pyyaml
argparse
pefile